Core: * add helping methods for manipulating unit's health and use it where applicable * fix some conversion warnings and cleanup code (formatting, CRLF, tabs to spaces)
--HG--
branch : trunk

+++ src/server/scripts/Kalimdor/RazorfenDowns/instance_razorfen_downs.cpp
@@ -41,7 +41,7 @@ public:
 
         uint32 m_auiEncounter[MAX_ENCOUNTER];
 
-        uint8 uiGongWaves;
+        uint16 uiGongWaves;
 
         std::string str_data;
