+++ src/server/scripts/Northrend/DraktharonKeep/instance_drak_tharon_keep.cpp
@@ -63,7 +63,7 @@ public:
         uint64 uiNovosCrystal3;
         uint64 uiNovosCrystal4;
 
-        uint8 m_auiEncounter[MAX_ENCOUNTER];
+        uint16 m_auiEncounter[MAX_ENCOUNTER];
 
         std::string str_data;
